requirements.txt
opencv-python
numpy
streamlit
pillow